#97 만화 및 편지 만화 생성로직 수정

+++ backend/src/main/java/tuneandmanner/wiselydiarybackend/diary/domain/repository/DiaryRepository.java
@@ -1,16 +1,16 @@
 package tuneandmanner.wiselydiarybackend.diary.domain.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
+import tuneandmanner.wiselydiarybackend.diary.domain.entity.Diary;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
-import tuneandmanner.wiselydiarybackend.diary.domain.entity.Diary;
-
 import java.time.LocalDateTime;
 import java.util.List;
 import java.util.Optional;
 
+
 @Repository
-public interface DiaryRepository extends JpaRepository<Diary, Integer> {
+public interface DiaryRepository extends JpaRepository<Diary, Long> {
     Optional<Diary> findByMemberIdAndCreatedAtBetween(String memberId, LocalDateTime start, LocalDateTime end);
 
     List<Diary> findByMemberIdAndCreatedAtBetweenAndDiaryStatus(String memberId, LocalDateTime startOfDay, LocalDateTime endOfDay, String exist);
